+++ packages/data-context/src/data/ProjectConfigIpc.ts
@@ -157,7 +157,16 @@ export class ProjectConfigIpc extends EventEmitter {
 
       let resolved = false
 
-      this._childProcess.on('error', (err) => {
+      this._childProcess.on('error', (err: NodeJS.ErrnoException) => {
+        if (err.code === 'EPIPE') {
+          debug('EPIPE error in loadConfig() of child process %s', err)
+
+          // @ts-ignore
+          resolve()
+
+          return
+        }
+
         debug('unhandled error in child process %s', err)
         this.handleChildProcessError(err, this, resolved, reject)
         reject(err)
@@ -229,7 +238,16 @@ export class ProjectConfigIpc extends EventEmitter {
     return new Promise((resolve, reject) => {
       let resolved = false
 
-      this._childProcess.on('error', (err) => {
+      this._childProcess.on('error', (err: NodeJS.ErrnoException) => {
+        if (err.code === 'EPIPE') {
+          debug('EPIPE error in registerSetupIpcHandlers() of child process %s', err)
+
+          // @ts-ignore
+          resolve()
+
+          return
+        }
+
         this.handleChildProcessError(err, this, resolved, reject)
         reject(err)
       })